+++ src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClientX/Handlers/HandlerRequestType.cs
@@ -1,0 +1,12 @@
+﻿namespace Microsoft.Data.SqlClientX.Handlers
+{
+
+    /// <summary>
+    /// The request type for the handler. This should be appended to contain any
+    /// other requests types as well.
+    /// </summary>
+    internal enum HandlerRequestType
+    {
+        ConnectionRequest
+    }
+}
